# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sudan Export B2B Export Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: active
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ── Companies ──────────────────────────────────────────────────────────
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [supplier, buyer, null]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}/verify:
    patch:
      operationId: verifyCompany
      tags: [companies]
      summary: Approve or reject company verification (admin)
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationAction"
      responses:
        "200":
          description: Company verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ── RFQs ───────────────────────────────────────────────────────────────
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for current user
      parameters:
        - name: rfqId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Notifications ──────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for current user/company
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Register a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - unit
        - isActive
        - createdAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
        - slug
        - unit
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Company ──────────────────────────────────────────────────────────
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [supplier, buyer]
        registrationNumber:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
          enum: [pending, approved, rejected]
        verifiedAt:
          type: ["string", "null"]
        ownerId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - country
        - verificationStatus
        - ownerId
        - createdAt
    CompanyInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [supplier, buyer]
        registrationNumber:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - name
        - type
        - country
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        registrationNumber:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
    VerificationAction:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
        reason:
          type: ["string", "null"]
      required:
        - action

    # ── RFQ ─────────────────────────────────────────────────────────────
...
    Message:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: ["integer", "null"]
        senderId:
          type: string
        senderName:
          type: ["string", "null"]
        recipientId:
          type: string
        recipientName:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        body:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - senderId
        - recipientId
        - body
        - isRead
        - createdAt
    MessageInput:
      type: object
      properties:
        rfqId:
          type: ["integer", "null"]
        recipientId:
          type: string
        subject:
          type: ["string", "null"]
        body:
          type: string
      required:
        - recipientId
        - body

    # ── Notification ─────────────────────────────────────────────────────
...
    Document:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        uploadedBy:
          type: string
        type:
          type: string
        name:
          type: string
        url:
          type: string
        mimeType:
          type: ["string", "null"]
        sizeBytes:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - uploadedBy
        - type
        - name
        - url
        - createdAt
    DocumentInput:
      type: object
      properties:
        companyId:
          type: ["integer", "null"]
        type:
          type: string
        name:
          type: string
        url:
          type: string
        mimeType:
          type: ["string", "null"]
        sizeBytes:
          type: ["integer", "null"]
      required:
        - type
        - name
        - url