# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sudan Export B2B Export Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyType } from './companyType';
import type { CompanyVerificationStatus } from './companyVerificationStatus';

export interface Company {
  id: number;
  name: string;
  type: CompanyType;
  /** @nullable */
  registrationNumber?: string | null;
  country: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  verificationStatus: CompanyVerificationStatus;
  /** @nullable */
  verifiedAt?: string | null;
  ownerId: string;
  createdAt: string;
}