# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sudan Export B2B Export Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  unit: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}